WIP hashtable
methods can take either a szNameBuf OR a lHashVal however only
one is actually required. If a lHashVal is provided the current
implementation is broken as it enumerates a linear search using
the szNameBuf TLBString.

Implement a HashMap and convert the 'typeinfos' array into one
such that it is indexed by a lHashVal. Methods that get passed
szNameBuf with lHashVal := 0 shall then need to calculate a
LHashValOfName to do lookups instead of the linear name str search.

Signed-off-by: Edward O'Callaghan <edward@antitrust.cc>